+++ CHANGELOG.md
@@ -2,8 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.0]
+
+### Changed
+- Kept only best performing compression strategies for Url and b85
+
 ## [0.0.1]
 
 ### Added
-- compression strategies
+- ~30 compression strategies
 - benchmarking